Cleaned code by implementing OOP structure for pipelines
