+++ pipelines/__init__.py
@@ -1,0 +1,5 @@
+from .base_pipeline import BasePipeline
+from .tpi_pipeline import TPIPipeline
+from .ascor_pipeline import ASCORPipeline
+
+__all__ = ['BasePipeline', 'TPIPipeline', 'ASCORPipeline']
